requirements.txt
openai
langchain
faiss-cpu
streamlit
tiktoken
python_dotenv